fix(canvas): preserve 64-bit IDs end to end

+++ src/lib/canvas/canvas-folders.js
@@ -63,10 +63,10 @@ function findCanvasFolder(userId, canvas) {
   return sql`
     SELECT note_id FROM app.notes
     WHERE user_id = ${userId}::uuid
-      AND canvas_course_id = ${canvasCourseId}::int
+      AND canvas_course_id = ${canvasCourseId}::bigint
       AND CASE
-        WHEN ${hasAssignment} THEN canvas_assignment_id = ${canvasAssignmentId ?? 0}::int
-        WHEN ${hasModule}     THEN canvas_module_id = ${canvasModuleId ?? 0}::int
+        WHEN ${hasAssignment} THEN canvas_assignment_id = ${canvasAssignmentId ?? 0}::bigint
+        WHEN ${hasModule}     THEN canvas_module_id = ${canvasModuleId ?? 0}::bigint
         ELSE canvas_module_id IS NULL AND canvas_assignment_id IS NULL
       END
       AND is_folder = true AND deleted_at IS NULL
